provide stdin/out/err readers, writers, etc
Previously, the Flow module assumed that the application using the RExec
module would be the client executing code on behalf of a remote server,
and therefore would always want to write to and read from the
same streams.  In the case where the RExec-using application has requested
that a service start, the application will instead be the server, so the
streams can no longer be assumed to be stdin/stdout as in the client
case.

This patch attempts to provide an interface for interacting with flows
on whichever stream is appropriate, for the use of both client and
server applications.

Co-authored-by: linse <sschirme@gmail.com>